requirements.txt
matplotlib
accelerate 
tiktoken 
einops 
scipy 
transformers_stream_generator==0.0.4 
peft 
deepspeed
autopep8==1.6.0
Flask==2.1.2
Werkzeug==2.2.2
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
Flask-SocketIO==5.2.0
opencc-python-reimplemented
line-bot-sdk
flask_cors
cryptography
chromadb==0.4.22
huggingface-hub==0.20.2
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.8
langchain-openai==0.0.2
pypdf==3.17.4
sentence-transformers==2.2.2
transformers==4.36.2
